Change LA init_actions -> init_wp_actions
The architecture of plugin stack library was changed to cooperate better with
WordPress' action and filter hooks. add_action() is used to trigger an
Lamosty_Actions action being dispatched instead of doing add_action() in
Lamosty_Stores.

Architecture is still far from being usable, because I haven't used it
on a larger WordPress plugin yet. Its API, classes and methods might
change in future so if you plan on using it in your projects, keep that
in mind.